add missing integration tests
CO-1050

Add some tests that were present in DirectoryEntityDAOTest, but not in
integration_tests.py.

Also, make sure we're actually testing what we say we're testing: when
searching by query, the directory api return a 200 response even if
there are 0 results. it's not enough to just test the status code, we
also need to look at the actual result.

Also, fix two tests that had the same method name, and tweak one test to
make the formatting match the other tests.

Also, fix the command line parsing.